pyproject.toml
[tool.isort]
profile = "black"

[build-system]
requires = ["setuptools>=58", "wheel"]
build-backend = "setuptools.build_meta"

[project]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: Apache Software License",
]
name = "py123d"
version = "0.1.0"
authors = [{ name = "Daniel Dauner", email = "daniel.dauner@gmail.com" }]
description = "A Library for Driving Datasets."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}
dependencies = [
    "geopandas",
    "matplotlib",
    "numpy",
    "opencv-python",
    "pandas",
    "Pillow",
    "psutil",
    "pyarrow",
    "pyogrio",
    "pyquaternion",
    "ray[default]",
    "scipy",
    "shapely>=2.0.0",
    "tqdm",
    "notebook",
    "hydra_colorlog",
    "hydra-core",
    "trimesh",
    "viser",
    "laspy[lazrs]",
    "DracoPy",
    "omegaconf",
    "typing-extensions",
    "requests",
    "cachetools", # Should be remove again, if functools are use
]

[project.scripts]
py123d-viser = "py123d.script.run_viser:main"
py123d-bokeh = "py123d.script.run_bokeh:main"
py123d-conversion = "py123d.script.run_conversion:main"

[project.optional-dependencies]
dev = [
    "pyright",
    "ruff",
    "pre-commit",
    "pytest",
    "pytest-cov",
]
docs = [
    "Sphinx<9",
    "sphinx-rtd-theme",
    "sphinx-autobuild",
    "sphinx-copybutton",
    "myst-parser",
    "furo",
    "autoclasstoc",
    "sphinx-autodoc-typehints",
    "sphinxcontrib-youtube",
    "sphinx-design",
]
nuplan = [
    "SQLAlchemy==1.4.27",
    "rasterio",
    "aioboto3",
    "retry",
    "cachetools",
]
nuscenes = [
    "nuscenes-devkit==1.2.0",
]
waymo = [
    "tensorflow-cpu>=2.13",
    "protobuf>=6.31",
]
ffmpeg = [
    "imageio[ffmpeg]",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
py123d = ["**/*.yaml", "**/*.yml", "**/*.proto", "**/*.xodr.gz"]

[project.urls]
"Homepage" = "https://github.com/autonomousvision/py123d"
"Bug Tracker" = "https://github.com/autonomousvision/py123d/issues"


[tool.ruff]
line-length = 120
lint.select = [
    "E",  # pycodestyle errors.
    "F",  # Pyflakes rules.
    "PLC",  # Pylint convention warnings.
    "PLE",  # Pylint errors.
    "PLR",  # Pylint refactor recommendations.
    "PLW",  # Pylint warnings.
    "I",  # Import sorting.
]
lint.ignore = [
    "E731",  # Do not assign a lambda expression, use a def.
    "E741",  # Ambiguous variable name. (l, O, or I)
    "E501",  # Line too long.
    "E402",  # Module level import not at top of file
    "PLR2004",  # Magic value used in comparison.
    "PLR0915",  # Too many statements.
    "PLR0913",  # Too many arguments.
    "PLC0414",  # Import alias does not rename variable. (this is used for exporting names)
    "PLC0415",  # Import should be at the top-level of a file.
    "PLC1901",  # Use falsey strings.
    "PLR5501",  # Use `elif` instead of `else if`.
    "PLR0911",  # Too many return statements.
    "PLR0912",  # Too many branches.
    "PLW0603",  # Global statement updates are discouraged.
    "PLW2901",  # For loop variable overwritten.
    "PLW0642",  # Reassigned self in instance method.
    "PLR0904",  # Checks for classes with too many public methods
    "PLR0917",  # Checks for function definitions that include too many positional arguments.
    "PLR0914",  # Checks for functions that include too many local variables.
    "PLR6301",  # Checks for the presence of unused self parameter in methods definitions.
    "PLC0206",  # Checks for dictionary iterations that extract the dictionary value via explicit indexing, instead of using .items().
]

exclude = [
    "__init__.py",
    "src/py123d/conversion/datasets/wod/waymo_open_dataset/protos",
]
fixable = ["ALL"]
unfixable = []

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
